Crate polkadot_runtime_metrics

Runtime metric interface similar to native Prometheus metrics.

This is intended to be used only for testing and debugging and must never be used in production. It requires the Substrate wasm tracing support and command line configuration: --tracing-targets wasm_tracing=trace.

Structs

Counter

A dummy Counter.

CounterVec

A dummy CounterVec.

Histogram

A dummy Histogram

Functions

get_current_time

Dummy implementation - always 0